mgx2r-package

Import mesh and cell graph files from MorphoGraphX software as tidy dataframes. Propose visualization solutions for mesh and cell graph files from MorphoGraphX with Plotly and Shiny.

Import from MorphoGraphX (MGX)

The function read_mgxPly imports meshes created with MorphoGraphX and saved as .ply as 3D mesh objects. The function read_mgxCellGraph imports cell graphs created with MorphoGraphX and saved as .ply as list of dataframes.

Visualisation and data exploration

Examples of simple data exploration of variables stored in the cell graphs are provided in the vignettes.

Export to MorphoGraphX (MGX)

The function mesh2ply exports a 3D mesh object from R to a .ply file that can be opened in MorphoGraphX (MGX).

Data

One example dataset is proposed with mgx2r. These data are meshes of plant tissues coming from the segmentation of confocal images.

General description
These data were generated by Marion Louveaux during her PhD thesis done at the ENS de Lyon between 2012 and 2015 under the supervision of Olivier Hamant. A shoot apical meristem of Arabidopsis thaliana, carrying a fluorescent membrane marker (LTi6B-GFP), was dissected and cut from the main stem of the plant, put in an in vitro media and imaged over 48h every 12h with a confocal microscope. The original name of the experiment is '2013-02-12_LTi6B_dis_A' (date of the experiment, name of fluorescent marker, preparation of the meristem -dissected- and name of the meristem).
Data content
Two versions of the same dataset are proposed here: the 'full' version contains the complete meshes and cell graphs of the five time points, while the cropped version contains cropped meshes of the first three time points only of the full dataset. Use the cropped dataset for quick tests. 'T0' to 'T48' refers to the time point of acquisition of the image after the start of the experiment in hours.
Data generation
For both full and cropped data. Mesh .ply and cell graph .ply were saved in MorphoGraphX following the instructions below:

Process -> Mesh -> Cell Mesh -> Convert to 2D cell tissue

Process -> Mesh -> Heatmap -> Heatmap then select geometry and area. This computes the area heatmap.

Process -> Mesh -> Heatmap -> Heatmap Export to Attr Map with custom name "Geometry/Area". This exports the calculated areas to the attributes.

Process -> Mesh -> Cell Axis -> Curvature -> Compute Tissue Curvature (with default parameters e.g. radius to 10 microns) This computes the curvature of the tissue.

Process -> Mesh -> Cell Axis -> Cell Axis Export to Attr Map with custom name "Cell axis curv10microns". This exports the calculated curvatures to the attributes.

Process -> Mesh -> Save Cell Graph Ply file saves both mesh .ply and cell graph .ply to the disk.

Meshes and cell graphs files were generated in MorphoGraphX (http://www.mpipz.mpg.de/MorphoGraphX) from the segmentation of these confocal images. Meshes contain information relative to the geometry of the tissue. Cell graphs contain data relative to the area of the cells and local curvature of the tissue (computed with a radius of 10 microns).